requirements.txt
tortoise-orm
aiosqlite
asyncpg
tornado
pyyaml
react
